1. one pcg interface supports ORG/TRANS poly_order=0/1; 2. delete old pcg trans files; 3. example type double

+++ examples/read_array.h
@@ -12,7 +12,7 @@
 #include <iostream>
 
 void readArrayFromFile(uint32_t size, const char *filename,
-                       float *matrix) {
+                       double *matrix) {
     FILE *myFile;
     myFile = fopen(filename, "r");
     if (myFile == NULL) {
@@ -21,7 +21,8 @@ void readArrayFromFile(uint32_t size, const char *filename,
     }
 
     for (uint32_t i = 0; i < size; i++) {
-        int ret = fscanf(myFile, "%f,", &matrix[i]);
+        int ret = fscanf(myFile, "%lf,", &matrix[i]); // for double
+//        int ret = fscanf(myFile, "%f,", &matrix[i]); // for float
     }
 
     fclose(myFile);
